refactor: revert domain directory creation and remove logging
