+++ packages/oasis-actions/src/helpers/ajna/AjnaPosition.ts
@@ -1,6 +1,6 @@
 import BigNumber from 'bignumber.js'
 
-import { IRiskRatio, RiskRatio } from '../../domain'
+import { IRiskRatio, RiskRatio } from '../calculations'
 import { IAjnaPosition, Pool } from '../../types/ajna'
 import { Address, AjnaError } from '../../types/common'
 import { ZERO } from '../constants'
