+++ packages/oasis-actions/src/index.ts
@@ -1,7 +1,7 @@
 export * as action from './actions'
 export { ActionFactory } from './actions/actionFactory'
-export { IBasePosition, IPosition, Position, Swap } from './domain/Position'
-export { IRiskRatio, RiskRatio } from './domain/RiskRatio'
+export { IBasePosition, IPosition, Position, Swap } from './helpers/calculations/Position'
+export { IRiskRatio, RiskRatio } from './helpers/calculations/RiskRatio'
 export * from './helpers/addresses'
 export * from './helpers/constants'
 export * as operations from './operations'
