+++ packages/oasis-actions/src/strategies/aave/open/open.ts
@@ -2,8 +2,8 @@ import BigNumber from 'bignumber.js'
 import { providers } from 'ethers'
 
 import { Unbox } from '../../../../../../helpers/types/common'
-import { IBaseSimulatedTransition, Position } from '../../../domain/Position'
-import { IRiskRatio } from '../../../domain/RiskRatio'
+import { IBaseSimulatedTransition, Position } from '../../../helpers/calculations/Position'
+import { IRiskRatio } from '../../../helpers/calculations/RiskRatio'
 import { amountFromWei, amountToWei, calculateFee } from '../../../helpers'
 import {
   DEFAULT_FEE,
